Add files via upload
Import this for use in DNN problems. 
One needs to define a loss_grad function inside the DNN class which must perform following functions: - 
accept a vector param, 
a boolean no_grad and,
inputs and labels for loss computation; 
assign param to  model parameters; 
set the gradients of model parameters to zero; 
if no_grad is False, then compute the loss, do backward pass and return a tuple of scalar loss value and 1D vector of gradients; 
if no_grad is True, the it should return just the scalar loss value,i.e. no need for backward pass.